Refactor: 기존 ItemService interface 삭제
